page: post: add cheatsheets
Cheatsheets are posts invisible from the front page in which I write things that I will find useful in the future. I often find myself searching for the same information over and over especially during my studies; I hope this will put an end to this. Cheatsheets page was made with the help of:

https://www.amitmerchant.com/how-to-categorize-your-posts-in-jekyll/